python_Flask/requirements.txt
python-telegram-bot == 13.7
Flask==2.0.1
requests==2.26.0